% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{The probability of shared birthdays}
\usage{
birthday(x)
}
\arguments{
\item{x}{a quantitative variable}
}
\value{
the probability of two or more people in a group of x having the same birthday
}
\description{
The probability of shared birthdays
}
\examples{
birthday(20:24)
}
