% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{Fish caught in the Tenessee River and its tributaries}
\format{
A data frame with 144 rows and 6 variabless:
\describe{
  \item{RIVER}{Three letter abbreviation of the RIver}
  \item{MILE}{Mile count for where the fish was caught}
  \item{SPECIES}{The species of the fish}
  \item{LENGTH}{The length of the fish}
  \item{WEIGHT}{THe weight of the fish}
  \item{DDT}{The amount of DDT found in the fish}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set with six variables, two of which are quantitative.
Interest is in the amount of DDT in their flesh
}
\keyword{datasets}
