% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function to Estimate Confidence Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of samples to create}

\item{x}{the sample to be replication from (could be considered the population)}

\item{fun}{the statistic function to calculate (must be usable in apply())}

\item{alpha}{the significance level for the confidence interval}

\item{cx}{the font size of the labels in the histogram}

\item{...}{additional arguments passed to the histogram}
}
\value{
a histogram showing the distribution of the samples and the population values and confidence interval for the given statistic function. Also a list containing ci, x, fun, and xstat
}
\description{
Bootstrap Function to Estimate Confidence Interval
}
\examples{
myboot2(10000, x = c(1, 1, 2, 2, 3, 3, 3, 3, 2, 4, 5, 5), fun = "mean",
alpha = 0.1, xlab = "mean", cx = 1.5)
}
