% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{CLT Uniform Distribution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{the size of each sample}

\item{iter}{the number of samples}

\item{a}{the minimum possible value for the data}

\item{b}{the maximum possible value for the data}
}
\value{
a histogram showing the Central Limit Theorem for the Uniform distribution
}
\description{
CLT Uniform Distribution
}
\examples{
\dontrun{mycltu(n = 3, iter = 10000, a = 2, b = 5)}
}
