% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{A Normal Distribution plot with shaded probability P(Y <= a)}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{any real number, the mean of the distribution}

\item{sigma}{any positive number, the standard deviation of the distribution}

\item{a}{any real number, the value of the desired probability}
}
\value{
a plot showing the distribution curve with the probability area shaded and a list of the mean, standard deviation, and area values
}
\description{
A Normal Distribution plot with shaded probability P(Y <= a)
}
\examples{
myncurve(2, 3, 1)
}
