% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Barplots of numbers picked from given samples}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{a positive number, the sample size}

\item{iter}{a positive number, the number of samples}

\item{time}{a positive integer, the time between each sample in seconds}
}
\value{
barplots showing the distribution of what numbers were picked from n trials from numbers 1-10 from iter number of samples. The barplots are separated by time seconds.
}
\description{
Barplots of numbers picked from given samples
}
\examples{
mysample(n = 1000, iter = 30)
mysample(n = 1000, iter = 1, time = 1)
}
