% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Tickets To Be Sold For a Flight}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats in a flight}

\item{gamma}{the probability a flight will be truly overbooked}

\item{p}{the probability of someone who booked a ticket actually showing up}
}
\value{
two plots showing the discrete and continuous cases of calculating how many tickets to sell and a list of the values
}
\description{
Calculate Tickets To Be Sold For a Flight
}
\examples{
ntickets(400, 0.02, 0.95)
}
