% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{A scatter plot with marginal histograms}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{A quantitative vector}

\item{y}{A quantitative vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A scatter plot of x vs y with margianl histograms
}
\description{
A scatter plot with marginal histograms
}
\examples{
scatterhist(x = 1:10, y = 11:20)
}
